+++ packages/bridge-status-controller/src/utils/history.ts
@@ -206,6 +206,7 @@ export const getInitialHistoryItem = (
     originalTransactionId,
     actionId,
     tokenSecurityTypeDestination,
+    inputPrimaryDenomination,
     batchSellData,
     quoteIds,
   } = args;
@@ -255,6 +256,9 @@ export const getInitialHistoryItem = (
     ...(activeAbTests && { activeAbTests }),
     ...(tokenSecurityTypeDestination !== undefined && {
       tokenSecurityTypeDestination,
+    }),
+    ...(inputPrimaryDenomination !== undefined && {
+      inputPrimaryDenomination,
     }),
   };